# Ordered execution of tasks per actor handle