--- a/src/ray/core_worker/transport/direct_actor_transport.cc
+++ b/src/ray/core_worker/transport/direct_actor_transport.cc
@@ -116,8 +116,8 @@ Status CoreWorkerDirectActorTaskSubmitter::SubscribeActorUpdates() {
 
 void CoreWorkerDirectActorTaskSubmitter::ConnectAndSendPendingTasks(
     const ActorID &actor_id, std::string ip_address, int port) {
-  std::unique_ptr<rpc::DirectActorClient> grpc_client(
-      new rpc::DirectActorClient(ip_address, port, client_call_manager_));
+  std::shared_ptr<rpc::DirectActorClient> grpc_client =
+      rpc::DirectActorClient::make(ip_address, port, client_call_manager_);
   RAY_CHECK(rpc_clients_.emplace(actor_id, std::move(grpc_client)).second);
 
   // Submit all pending requests.