--- a/src/ray/protobuf/direct_actor.proto
+++ b/src/ray/protobuf/direct_actor.proto
@@ -16,9 +16,17 @@ message ReturnObject {
 message PushTaskRequest {
   // The task to be pushed.
   TaskSpec task_spec = 1;
-  // The sequence number of the task for this client.
+  // The sequence number of the task for this client. This must increase
+  // sequentially starting from zero for each actor handle. The server
+  // will guarantee tasks execute in this sequence, waiting for any
+  // out-of-order request messages to arrive as necessary.
   int64 sequence_number = 2;
-  // The max sequence number the client has processed responses for.
+  // The max sequence number the client has processed responses for. This
+  // is used to handle faults. The server will cancel and skip any
+  // any PushTaskRequests with seqno <= this value. This means if a client
+  // fails to send a request for whatever reason, it just needs to bump
+  // client_processed_up_to be greater than that request's seqno to avoid
+  // the server waiting forever for that missing message.
   int64 client_processed_up_to = 3;
 }