@zhijunfu 2019-09-09
Thanks, have you got a chance to test the ordering and performance?