@ericl 2019-09-09
Which tests should I run for performance? The ordering is fine.

On Sun, Sep 8, 2019, 10:51 PM Zhijun Fu <notifications@github.com> wrote:

> Thanks, have you got a chance to test the ordering and performance?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/ray-project/ray/pull/5664?email_source=notifications&email_token=AAADUSTIJ6TBD4TAUW5MBIDQIXP4LA5CNFSM4IUVLOJ2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD6GKBHI#issuecomment-529309853>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAADUSWBFAQ6V63MMPWT2RLQIXP4LANCNFSM4IUVLOJQ>
> .
>