@zhijunfu 2019-09-09
There are a couple of direct actor call performance tests in the asio rpc PR, which cover single node and cross nodes scenarios, and also for tasks with return and without return. You could leverage these tests for perf testing.

https://github.com/ray-project/ray/pull/5617/files#diff-3f030dc008041525a014ff6e73af7d64

src/ray/core_worker/test/core_worker_test.cc

```
TEST_F(SingleNodeTest, TestDirectActorTaskLocalPerformance) {
TEST_F(TwoNodeTest, TestDirectActorTaskCrossNodesPerformance) {
TEST_F(SingleNodeTest, TestDirectActorTaskLocalNoReturnPerformance) {
TEST_F(TwoNodeTest, TestDirectActorTaskCrossNodesNoReturnPerformance) {
```



> Which tests should I run for performance? The ordering is fine.
> […](#)
> On Sun, Sep 8, 2019, 10:51 PM Zhijun Fu ***@***.***> wrote: Thanks, have you got a chance to test the ordering and performance? — You are receiving this because you authored the thread. Reply to this email directly, view it on GitHub <#5664?email_source=notifications&email_token=AAADUSTIJ6TBD4TAUW5MBIDQIXP4LA5CNFSM4IUVLOJ2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD6GKBHI#issuecomment-529309853>, or mute the thread <https://github.com/notifications/unsubscribe-auth/AAADUSWBFAQ6V63MMPWT2RLQIXP4LANCNFSM4IUVLOJQ> .