@ericl 2019-09-09
Here's some perf numbers:

```
Before this PR:
start submitting 100000 tasks
finish submitting 100000 tasks, which takes 1187 ms
finish executing 100000 tasks, which takes 1959 ms
```

```
After this PR:
start submitting 100000 tasks
finish submitting 100000 tasks, which takes 1444 ms  <--- slightly slower
                                                          due to admission control
finish executing 100000 tasks, which takes 1649 ms  <--- slightly faster
```


Another improvement is that now we can submit 1 million tasks without crashing. Before this PR, the actor would die if you tried to benchmark that, presumably due to lack of backpressure.