@zhijunfu 2019-09-11
Thanks. In addition to the comments above, I also have some high level comments:

On the performance side, could you please run the test cases I mentioned above, and compare with asio rpc? The test cases are `TestDirectActorTaskLocalPerformance` and `TestDirectActorTaskLocalNoReturnPerformance`.

In my previous testing, asio RPC is 30% faster than GRPC for tasks with return objects, and 100% faster for tasks with no return. 

It's reported that GRPC stream can be faster than GRPC unary as it doesn't require a reply for each request. Wondering if you have plans to support GRPC stream for direct call? 

**Asio RPC**

[TestDirectActorTaskLocalPerformance]
```
I0823 21:38:32.942021 83557 core_worker_test.cc:401] finish submitting 10000 tasks, which takes 206 ms
I0823 21:38:33.062526 83557 core_worker_test.cc:410] finish executing 10000 tasks, which takes 326 ms
```

[TestDirectActorTaskLocalNoReturnPerformance]
```
I0823 21:38:33.698083 83557 core_worker_test.cc:401] finish submitting 10000 tasks, which takes 190 ms
I0823 21:38:33.698366 83557 core_worker_test.cc:410] finish executing 10000 tasks, which takes 190 ms
```

**GRPC unary**

[TestDirectActorTaskLocalPerformance]
```
I0823 21:49:33.739959 104885 core_worker_test.cc:401] finish submitting 10000 tasks, which takes 350 ms
I0823 21:49:33.818540 104885 core_worker_test.cc:410] finish executing 10000 tasks, which takes 429 ms
```

[TestDirectActorTaskLocalNoReturnPerformance]
```
I0823 21:49:34.596382 104885 core_worker_test.cc:401] finish submitting 10000 tasks, which takes 331 ms
I0823 21:49:34.667493 104885 core_worker_test.cc:410] finish executing 10000 tasks, which takes 402 ms
```