@ericl 2019-09-11
@zhijunfu I didn't run those tests since they are non-trivial to rebase on this PR. In any case, we just can't be using a different RPC layer for direct actor calls. It's much more important for Ray to have all traffic go over the same protocol, and we can consider optimizations like application-level batching in the future.

Feel free to try to rebase those tests on this PR, but even if using raw TCP faster it doesn't mean we can just ditch gRPC.