@zhijunfu 2019-09-11
Just to be clear: I'm not against using GRPC, I agree it's important to have all traffic to go through the same RPC, and it would be nice if we can use GRPC. I just want to make sure it's robust (on request ordering), and has comparable performance with asio (not too much regressions). 

> @zhijunfu I didn't run those tests since they are non-trivial to rebase on this PR. In any case, we just can't be using a different RPC layer for direct actor calls. It's much more important for Ray to have all traffic go over the same protocol, and we can consider optimizations like application-level batching in the future.
> 
> Feel free to try to rebase those tests on this PR, but even if using raw TCP faster it doesn't mean we can just ditch gRPC.