@ericl 2019-09-11
Btw, maybe you missed the benchmark above, which shows the current implementation is slightly faster than the previous out-of-order gRPC implementation, at ~80 thousand calls/s.