@ericl 2019-09-11
Thanks @zhijunfu , I added a magic token that would tell mock worker to check the sequence of messages. Let me know if you had something better in mind.